fix: ошибки по всему репозиторию

+++ src/main/java/org/gfs/glowcore/core/util/math/math.java
@@ -1,6 +1,6 @@
 package org.gfs.glowcore.core.util.math;
 
-
+import static org.lwjgl.glfw.GLFW.glfwGetTime;
 
 
 public class math {
@@ -11,11 +11,12 @@ public class math {
     public float z;
     public float w;
 
+
     
     public float timeDelta() {
         float deltaTime = 0.0f;
         float lastFrame = 0.0f;
-        float currentFrame = glfwGetTime();
+        float currentFrame = (float) glfwGetTime();
         deltaTime = currentFrame-lastFrame;
         lastFrame = currentFrame;
